refactor: 이름 원시값 포장
