+++ src/main/java/domain/Player.java
@@ -1,20 +1,13 @@
 package domain;
 
 public class Player {
-    private final String name;
+    private final Name name;
 
     public Player(String name) {
-        validateName(name);
-        this.name = name;
-    }
-
-    private void validateName(String name) {
-        if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException("이름은 공백일 수 없습니다.");
-        }
+        this.name = new Name(name);
     }
 
     public String getName() {
-        return name;
+        return name.getValue();
     }
 }
